rename service names
