+++ web/src/services/delete-project.ts
@@ -2,39 +2,34 @@ import { useMutation, useQueryClient } from '@tanstack/react-query';
 import useSupabaseBrowser from '@/lib/supabase/supabase-client';
 import { TypedSupabaseClient } from '@/lib/supabase/types';
 import { Database } from '@/utils/database.types';
-import { getQueryKey } from './_queryKeys';
+import { getQueryKey } from './_query-keys';
 
-type UpdateOptions = Database['public']['Tables']['Project']['Update'];
+type DeleteProjectInput = {
+  id: Database['public']['Tables']['Project']['Row']['id'];
+};
 
-type UpdateProjectInput = {
-  id: string;
-} & UpdateOptions;
-
-async function updateProject(
+async function deleteProject(
   client: TypedSupabaseClient,
-  inputValues: UpdateProjectInput
+  inputValues: DeleteProjectInput
 ) {
   const { data, error } = await client
     .from('Project')
-    .update({
-      name: inputValues.name,
-      color: inputValues.color,
-    })
+    .delete()
     .eq('id', inputValues.id)
     .select();
 
-  if (error) throw new Error('Error updating project');
+  if (error) throw new Error('Error deleting project');
 
   return data;
 }
 
-export const useUpdateProject = () => {
+export const useDeleteProject = () => {
   const client = useSupabaseBrowser();
   const queryClient = useQueryClient();
 
   return useMutation({
-    mutationFn: (inputValues: UpdateProjectInput) =>
-      updateProject(client, inputValues),
+    mutationFn: (inputValues: DeleteProjectInput) =>
+      deleteProject(client, inputValues),
     onSuccess: () => {
       queryClient.invalidateQueries({
         queryKey: getQueryKey('projects', 'list'),
